add gloabl hook
